refactor(http): Move GraphQL into http.NewRelicClient as Query()
